Add react error handling
