+++ web/app/js/components/App/index.tsx
@@ -1,97 +1,27 @@
-import React from "react";
-import { ApolloProvider } from "react-apollo";
-import { createMuiTheme, MuiThemeProvider } from "@material-ui/core/styles";
-import CssBaseline from "@material-ui/core/CssBaseline";
-import { BrowserRouter, Route, Switch } from "react-router-dom";
-import getApolloClient from "../../services/apollo";
-import {
-  ErrorPage,
-  NewsPage,
-  ReadingListPage,
-  FavoritesPage,
-  SyndicationPage,
-  AccountPage
-} from "../Page";
-import uiTheme from "../ui/theme";
-import Loader from "../ui/Loader";
-import InitQuery, { Data } from "../apollo/Query/Init";
-import { AppContext, UserContext } from "../context";
+import React, { Component, ErrorInfo } from "react";
+import App from "./App";
+import ErrorMessage from "./Error";
 
-function canBoostrap(data: Data | undefined): boolean {
-  if (!data) {
-    return false;
-  }
-  if (!data.users) {
-    return false;
-  }
-  if (!data.users.loggedIn) {
-    return false;
-  }
-  if (!data.app) {
-    return false;
-  }
-  if (!data.app.info) {
-    return false;
-  }
-  return true;
+interface State {
+  error: Error | null;
+  info: ErrorInfo | null;
 }
 
-export default function App(): JSX.Element {
-  const client = getApolloClient();
-  const theme = createMuiTheme(uiTheme);
+export default class Bootstrap extends Component<{}, State> {
+  state = {
+    error: null,
+    info: null
+  };
 
-  console.log(theme);
+  componentDidCatch(error: Error, info: ErrorInfo): void {
+    this.setState({
+      error: error,
+      info: info
+    });
+  }
 
-  return (
-    <ApolloProvider client={client}>
-      <CssBaseline />
-      <MuiThemeProvider theme={theme}>
-        <BrowserRouter>
-          <InitQuery>
-            {({ data, loading, error }) => {
-              if (loading) {
-                return <Loader />;
-              }
-
-              if (error) {
-                return <ErrorPage error={error} />;
-              }
-
-              if (!canBoostrap(data)) {
-                return null;
-              }
-
-              const { users, app } = data as Data;
-
-              return (
-                <AppContext.Provider value={app.info}>
-                  <UserContext.Provider value={users.loggedIn}>
-                    <Switch>
-                      <Route exact path="/" component={NewsPage} />
-                      <Route
-                        exact
-                        path="/reading-list"
-                        component={ReadingListPage}
-                      />
-                      <Route
-                        exact
-                        path="/favorites"
-                        component={FavoritesPage}
-                      />
-                      <Route
-                        exact
-                        path="/syndication"
-                        component={SyndicationPage}
-                      />
-                      <Route exact path="/account" component={AccountPage} />
-                    </Switch>
-                  </UserContext.Provider>
-                </AppContext.Provider>
-              );
-            }}
-          </InitQuery>
-        </BrowserRouter>
-      </MuiThemeProvider>
-    </ApolloProvider>
-  );
+  render(): JSX.Element {
+    const { info, error } = this.state;
+    return info ? <ErrorMessage info={info} error={error} /> : <App />;
+  }
 }
